+++ internal/server/main_test.go
@@ -1,0 +1,13 @@
+package server
+
+import (
+	"os"
+	"testing"
+
+	"memkv/internal/logger"
+)
+
+func TestMain(m *testing.M) {
+	logger.SetDefault(logger.Discard())
+	os.Exit(m.Run())
+}